B2B Catalog Permissions  (#695)
Added support for Magento Commerce (EE) edition features: Catalog Permissions and B2B Shared Catalog. This allows you to limit visibility of the catalog by categories and products based on customer_group_id.

In order to be compatible with Magento OS edition, we used observers to call factory method to create classes through the objectManger that only exists in Commerce. These observers are in charge of adding catalog_permissions attributes to _products and _categories indexes. We have one frontend observer that adds a filter to the generated secure API key that filters if attribute != 0.